# project/server/docs/swagger/htmlSwag/swagger.json
{
  "basePath" : "/",
  "definitions" : {
...
    "DataPJournal" : {
      "type" : "object",
      "properties" : {
        "day" : {
          "type" : "string"
        },
        "group" : {
          "format" : "int64",
          "type" : "integer"
        },
        "homework" : {
          "type" : "string"
        },
        "kid" : {
          "format" : "int64",
          "type" : "integer"
        },
        "mark" : {
          "type" : "string"
        },
        "per" : {
          "format" : "int64",
          "type" : "integer"
        },
        "predm" : {
          "type" : "string"
        },
        "style" : {
          "type" : "string"
        },
        "uuid" : {
          "type" : "string"
        },
        "weight" : {
          "format" : "int32",
          "type" : "integer"
        }
      }
    },
...
    "JsonArray" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "empty" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonNull" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonObject" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "empty" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        }
      }
    },
    "JsonPrimitive" : {
      "type" : "object",
      "properties" : {
        "asBigDecimal" : {
          "type" : "number"
        },
        "asBigInteger" : {
          "type" : "integer"
        },
        "asBoolean" : {
          "type" : "boolean"
        },
        "asByte" : {
          "format" : "byte",
          "type" : "string"
        },
        "asCharacter" : {
          "type" : "string"
        },
        "asDouble" : {
          "format" : "double",
          "type" : "number"
        },
        "asFloat" : {
          "format" : "float",
          "type" : "number"
        },
        "asInt" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asJsonArray" : {
          "$ref" : "#/definitions/JsonArray"
        },
        "asJsonNull" : {
          "$ref" : "#/definitions/JsonNull"
        },
        "asJsonObject" : {
          "$ref" : "#/definitions/JsonObject"
        },
        "asJsonPrimitive" : {
          "$ref" : "#/definitions/JsonPrimitive"
        },
        "asLong" : {
          "format" : "int64",
          "type" : "integer"
        },
        "asNumber" : {
          "$ref" : "#/definitions/Number"
        },
        "asShort" : {
          "format" : "int32",
          "type" : "integer"
        },
        "asString" : {
          "type" : "string"
        },
        "boolean" : {
          "type" : "boolean"
        },
        "jsonArray" : {
          "type" : "boolean"
        },
        "jsonNull" : {
          "type" : "boolean"
        },
        "jsonObject" : {
          "type" : "boolean"
        },
        "jsonPrimitive" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "boolean"
        },
        "string" : {
          "type" : "boolean"
        }
      }
    },
    "Number" : {
      "type" : "object"
    },
...
  },
  "host" : "localhost:8080",
  "info" : {
    "description" : "This is a sample.",
    "title" : "EJournal",
    "version" : "v1"
  },
  "paths" : {
...
    "/pjournal/addHomework" : {
      "post" : {
        "operationId" : "addHomework",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataPJournal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/pjournal/addMark" : {
      "post" : {
        "operationId" : "addMark",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataPJournal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/pjournal/getInfoP1" : {
      "post" : {
        "operationId" : "getInfoP1",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataPJournal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/pjournal/getInfoP2" : {
      "post" : {
        "operationId" : "getInfoP2",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataPJournal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
    "/pjournal/getInfoP3" : {
      "post" : {
        "operationId" : "getInfoP3",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/DataPJournal"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/JsonObject"
            }
          }
        }
      }
    },
...
  },
  "schemes" : [ "http" ],
  "swagger" : "2.0",
...
}